+++ src/react-router-6.ts
@@ -10,6 +10,6 @@
  * See the License for the specific language governing permissions and limitations under the License.
  */
 
-import AuthRequired from './AuthRequired';
+import SecureOutlet from './SecureOutlet';
 
-export { AuthRequired };
+export { SecureOutlet };