+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ### Features
 
 - [#275](https://github.com/okta/okta-react/pull/275)
-  - Adds new component `<AuthRequired>` for integration with `react-router 6.x`. It should be imported from `@okta/okta-react/react-router-6`
+  - Adds new component `<SecureOutlet>` for integration with `react-router 6.x`. It should be imported from `@okta/okta-react/react-router-6`
 
 # 6.8.0